requirements.txt
aioboto3
aiofiles
bokeh==2.4.3  # Used in the nuBoard dashboard
boto3==1.24.59  # Used in s3_store.py to download data from S3.
cachetools  # Used widely
casadi  # Used for optimization solving
control==0.9.1  # Used for LQR controller synthesis
coverage  # Used in computing code coverage
docker  # To run remote submissions
Fiona  # Used in GpkgMapsDB.py
geopandas>=0.12.1  # Used to read maps
#grpcio==1.43.0  # Used for communication with submission container
#grpcio-tools==1.43.0  # Used for communication with submission container
guppy3==3.1.2
hydra-core==1.1.0rc1  # Used for configs
hypothesis
joblib
jupyter  # Used to show the notebook
jupyterlab
matplotlib  # Used for rendering
mock
moto
nest_asyncio  # Used to nest event loops when running nuBoard/jupyter
numpy==1.23.4
opencv-python<=4.5.1.48  # Used widely
pandas    # Used widely
Pillow    # Used widely to render images
pre-commit  # Used in pre-commit hooks
psutil  # Used widely to get the number of CPUs
pyarrow # For parquet
pyinstrument    # Used widely as profiler
pyogrio # A performant backend for geopandas
pyquaternion>=0.9.5  # Used widely, avoid known bug with quaternion slerp
pytest    # Used widely
rasterio  # Used in GpkgMapsDB.py
ray # Used widely
requests  # Used in BlobStore classes
retry
rtree # Used in occupancy maps
s3fs # Used by pandas to read from/write to s3
scipy  # Used widely
selenium # Used in bokeh export png
setuptools==59.5.0  # Used in setup.py, pinned to not break pytorch
Shapely>=2.0.0  # Used widely
SQLAlchemy==1.4.27  # older versions don't work with some table definitions
sympy  # Use for symbolic algebra
testbook  # Used in testing jupyter notebooks
tornado  # Used in nuboard.py
tqdm  # Used widely
typer # Used for cli
ujson  # Used in serialiation_callback.py
urllib3  # Used in s3_store.py
